+++ src/Dap.Platform/Types.fs
@@ -1,69 +1,130 @@
 [<AutoOpen>]
-module Dap.Platform.Types
+module Dap.Platform.Types'
 
 open Dap.Prelude
 
-[<Measure>]
-type ms
+type QuitStats = {
+    ForceQuit : bool
+    ProcessTime : Instant
+    QuitDuration : Duration
+}
 
-type Scope = string
-type Kind = string
-type Key = string
+type StopStats = {
+    ForceStop : bool
+    ProcessTime : Instant
+    StopDuration : Duration
+}
 
-type Ident = {
+type IPlatform =
+    abstract Start : IRunnable<'runner, 'args, 'model, 'msg> -> unit
+
+type IEnv =
+    inherit IRunner<IEnv>
+    inherit IHandler<EnvReq>
+    inherit IAsyncHandler<EnvReq>
+    abstract Platform : IPlatform with get
+    abstract Logging : ILogging with get
+    abstract Scope : Scope with get
+    abstract State : EnvModel option with get
+
+and Spawner =
+    Ident -> IAgent
+
+and EnvModel = {
+    Services : Map<Kind, IAgent>
+    Spawners : Map<Kind, Spawner>
+    Agents : Map<Kind, Map<Key, IAgent>>
+}
+
+and EnvReq =
+    | DoQuit of bool * Callback<QuitStats>                  // forceQuit -> willQuit
+    | DoAddService of IAgent * Callback<int>                // -> servicesCount
+    | DoGetService of Kind * Callback<IAgent>               // -> service
+    | DoRegister of Kind * Spawner * Callback<int>          // -> spawnersCount
+    | DoGetAgent of Kind * Key * Callback<IAgent * bool>    // -> (agent, isNew)
+
+and EnvEvt =
+    | OnQuit of QuitStats
+    | OnNewAgent of Kind * Key * IAgent
+
+and EnvMsg =
+    | EnvReq of EnvReq
+    | EnvEvt of EnvEvt
+with interface IMsg
+
+and EnvOperate =
+        Operate<IEnv, EnvModel, EnvMsg>
+
+and EnvParam = {
+    Platform : IPlatform
+    Logging : ILogging
     Scope : Scope
+    Clock : IClock
+    GetSlowCap : GetSlowCap option
+}
+
+and IAgent =
+    inherit IRunner<IAgent>
+    inherit IHandler<AgentReq>
+    inherit IAsyncHandler<AgentReq>
+    //inherit IChannel<AgentEvt>
+    abstract Env : IEnv with get
+    abstract Ident : Ident with get
+
+and IAgent<'req, 'evt> =
+    inherit IAgent
+    inherit IPoster<'req>
+    inherit IAsyncPoster<'req>
+    abstract Actor : IActor<'req, 'evt> with get
+
+and IAgent<'model, 'req, 'evt> =
+    inherit IAgent<'req, 'evt>
+    abstract Actor : IActor<'model, 'req, 'evt> with get
+
+and IAsyncAgent<'model, 'req, 'evt> =
+    inherit IHandler<'req>
+
+and AgentSpec<'args, 'model, 'msg, 'req, 'evt> = {
+    Actor : ActorSpec<IAgent, 'args, 'model, 'msg, 'req, 'evt>
+    OnAgentEvent : OnAgentEvent<'model> option
+    GetSlowCap : GetSlowCap option
+}
+
+and OnAgentEvent<'model> = IAgent -> 'model -> AgentEvt -> unit
+
+and AgentParam = {
+    Env : IEnv
     Kind : Kind
     Key : Key
 }
 
-type IMsg = interface end
-
-type Response<'req, 'res> =
-    | Ack of 'req * 'res
-    | Nak of 'req * string * obj
-
-type Reply<'res> = Response<IMsg, 'res>
-
-exception ReplyException of string * obj
-
-type Callback<'res> = (Reply<'res> -> unit) option
-
-type IHandler<'req> =
-    abstract Handle : 'req -> unit
-
-type IPoster<'msg> =
-    abstract Post : 'msg -> unit
-
-and IChannel<'evt> =
-    abstract OnEvent : IEvent<'evt> with get
-
-and IActor =
-    inherit ILogger
-    abstract Ident : Ident with get
-
-and IActor<'req, 'evt> =
-    inherit IActor
-    inherit IHandler<'req>
-    inherit IChannel<'evt>
-
-and IActor<'model, 'req, 'evt> =
-    inherit IActor<'req, 'evt>
-    abstract State : 'model option with get
-
-and ActorSpec<'runner, 'args, 'model, 'msg, 'req, 'evt> = {
-    Logic : Logic<'runner, 'args, 'model, 'msg>
-    NewArgs : unit -> 'args
-    WrapReq : Wrapper<'msg, 'req>
-    GetOnEvent : 'model -> IEvent<'evt>
+and AgentModel<'args, 'model, 'msg, 'req, 'evt> = {
+    Spec : AgentSpec<'args, 'model, 'msg, 'req, 'evt>
+    Actor : 'model
 }
 
-type NoArgs = NoArgs
+and AgentReq =
+    | DoStop of bool * Callback<StopStats>           // forceStop
 
-and NoModel = NoModel
+and AgentEvt =
+    | OnWillStop of Callback<unit> 
+    | OnDidStop of StopStats
 
-and NoMsg = NoMsg
+and AgentMsg<'args, 'model, 'msg, 'req, 'evt> =
+    | AgentReq of AgentReq
+    | AgentEvt of AgentEvt
+    | ActorMsg of 'msg
+    | ActorMsg' of AgentOperate<'args, 'model, 'msg, 'req, 'evt>
 with interface IMsg
 
-and NoReq = NoReq
+and AgentOperate<'args, 'model, 'msg, 'req, 'evt> =
+    Operate<IAgent,
+            AgentModel<'args, 'model, 'msg, 'req, 'evt>,
+            AgentMsg<'args, 'model, 'msg, 'req, 'evt>>
 
-and NoEvt = NoEvt
+let identOf (scope : Scope) (kind : Kind) (key : Key) : Ident =
+    {
+        Scope = scope
+        Kind = kind
+        Key = key
+    }