+++ src/Dap.Prelude/Union.fs
@@ -1,3 +1,0 @@
-[<AutoOpen>]
-[<RequireQualifiedAccess>]
-module Dap.Prelude.Union